adjusted for refactor
